// apps/web/src/lib/community/schema.ts
export type CommunityMetadataDraft = {
  name: string;
  symbol: string;
  description: string;
  externalLinkLabel: string;
  externalLinkUrl: string;
};